+++ src/causalitytests/tests_entropy_based/ExactSimplexIntersectionTest.jl
@@ -5,7 +5,8 @@ import CausalityToolsBase: TriangulationBinning, ExactIntersection
 
 """
     ExactSimplexIntersectionTest(k::Int = 1, l::Int = 1, m::Int = 1, n::Int = 1,
-        τ::Int = 1, estimator::TransferEntropyEstimator = TransferOperatorGrid(b = 2),
+        τ::Int = 1, 
+        estimator::BinningTransferEntropyEstimator = TransferOperatorGrid(b = 2),
         n_pts::Int = 10000, binning_summary_statistic::Function = StatsBase.mean,
         binning::Union{RectangularBinning, Vector{RectangularBinning}},
         ηs::Union{AbstractVector{Int}})
@@ -107,7 +108,7 @@ Base.@kwdef mutable struct ExactSimplexIntersectionTest{N} <: TransferEntropyCau
     The transfer entropy estimator used to estimate transfer entropy *after* the invariant measure
     over the triangulated delay reconstruction has been estimated.
     """
-    estimator::TransferEntropyEstimator = TransferOperatorGrid(b = 2)
+    estimator::BinningTransferEntropyEstimator = TransferOperatorGrid(b = 2)
 
     """ The number of points to generate from the invariant distribution over the triangulation. """
     n_pts::Int = 10000
@@ -132,7 +133,7 @@ Base.@kwdef mutable struct ExactSimplexIntersectionTest{N} <: TransferEntropyCau
             estimator::E, n_pts::Int, 
             binning_summary_statistic::Function, 
             binning::Union{RectangularBinning, Vector{RectangularBinning}}, 
-            ηs) where {E <: TransferEntropyEstimator}
+            ηs) where {E <: BinningTransferEntropyEstimator}
 
         N = length(ηs) # length of return vector when used with `causality`
         return new{N}(k, l, m, n, τ, estimator, n_pts,