+++ src/causalitytests/tests_entropy_based/VisitationFrequencyTest.jl
@@ -199,10 +199,15 @@ end
 
 
 function causality(source::AbstractVector{T}, target::AbstractVector{T}, p::VisitationFrequencyTest) where {T <: Real}
-    [p.binning_summary_statistic(
-        transferentropy(source, target, p.binning, 
-            p.k, p.l, p.m, η = η, τ = p.τ, 
-            estimator = p.estimator)) for η in p.ηs]
+    [transferentropy(source, target, p.k, p.l, p.m, p.binning, η = η, τ = p.τ, 
+        summary_statistic = p.binning_summary_statistic, 
+            estimator = p.estimator) for η in p.ηs]
+end
+
+function causality(source::AbstractVector{T}, target::AbstractVector{T}, cond::AbstractVector{T}, p::VisitationFrequencyTest) where {T <: Real}
+    [transferentropy(source, target, cond, p.k, p.l, p.m, p.n, p.binning, η = η, τ = p.τ, 
+        summary_statistic = p.binning_summary_statistic, 
+            estimator = p.estimator) for η in p.ηs]
 end